Unify the remove/place code
